WIP: update to latest libgit2 API

+++ ObjectiveGit/GTFilterList.h
@@ -13,7 +13,7 @@
 @class GTRepository;
 
 /// The options for loading a filter list. See libgit2 for more information.
-typedef NS_OPTIONS(NSInteger, GTFilterListOptions) {
+typedef NS_OPTIONS(uint32_t, GTFilterListOptions) {
 	GTFilterListOptionsDefault = GIT_FILTER_DEFAULT,
 	GTFilterListOptionsAllowUnsafe = GIT_FILTER_ALLOW_UNSAFE,
 };
